# Support pushing dereferences within lambdas into table scan