--- a/plugin/trino-hive/src/main/java/io/trino/plugin/hive/HivePageSourceProvider.java
+++ b/plugin/trino-hive/src/main/java/io/trino/plugin/hive/HivePageSourceProvider.java
@@ -494,7 +494,8 @@ public class HivePageSourceProvider
                                     projectedColumn.getDereferenceIndices(),
                                     projectedColumn.getDereferenceNames(),
                                     fromHiveType,
-                                    createTypeFromCoercer(typeManager, fromHiveType, columnHandle.getHiveType(), coercionContext));
+                                    createTypeFromCoercer(typeManager, fromHiveType, columnHandle.getHiveType(), coercionContext),
+                                    projectedColumn.getSubfields());
                         });
 
                         return new HiveColumnHandle(