… core/trino-main/src/main/java/io/trino/sql/planner/iterative/rule/PushDownFieldReferenceLambdaThroughProject.java
… core/trino-main/src/main/java/io/trino/sql/planner/iterative/rule/PushFieldReferenceLambdaIntoTableScan.java
… core/trino-main/src/main/java/io/trino/sql/planner/iterative/rule/PushFieldReferenceLambdaThroughFilterIntoTableScan.java
… core/trino-main/src/test/java/io/trino/sql/analyzer/TestFeaturesConfig.java
… core/trino-main/src/test/java/io/trino/sql/planner/TestConnectorExpressionTranslator.java
… core/trino-main/src/test/java/io/trino/sql/planner/TestPartialTranslator.java
… core/trino-main/src/test/java/io/trino/sql/planner/assertions/ExpressionMatcher.java
… core/trino-main/src/test/java/io/trino/sql/planner/assertions/PlanMatchPattern.java
… core/trino-main/src/test/java/io/trino/sql/planner/iterative/rule/TestPushDownSubscriptLambdaRules.java
… core/trino-spi/pom.xml
… core/trino-spi/src/main/java/io/trino/spi/expression/ArrayFieldDereference.java
… lib/trino-plugin-toolkit/src/main/java/io/trino/plugin/base/projection/ApplyProjectionUtil.java
… lib/trino-plugin-toolkit/src/main/java/io/trino/plugin/base/subfield/Subfield.java
… lib/trino-plugin-toolkit/src/main/java/io/trino/plugin/base/subfield/SubfieldTokenizer.java
… lib/trino-plugin-toolkit/src/test/java/io/trino/plugin/base/subfield/TestSubfieldTokenizer.java
… plugin/trino-delta-lake/src/main/java/io/trino/plugin/deltalake/transactionlog/checkpoint/CheckpointEntryIterator.java
… plugin/trino-hive/src/main/java/io/trino/plugin/hive/HiveColumnProjectionInfo.java
… plugin/trino-hive/src/main/java/io/trino/plugin/hive/HiveMetadata.java
… plugin/trino-hive/src/main/java/io/trino/plugin/hive/parquet/ParquetPageSourceFactory.java
… plugin/trino-hive/src/main/java/io/trino/plugin/hive/util/HiveTypeUtil.java
… plugin/trino-hive/src/test/java/io/trino/plugin/hive/TestHiveColumnHandle.java
… plugin/trino-hive/src/test/java/io/trino/plugin/hive/TestHiveFileFormats.java
… plugin/trino-hive/src/test/java/io/trino/plugin/hive/TestHiveReaderProjectionsUtil.java
… plugin/trino-hive/src/test/java/io/trino/plugin/hive/optimizer/TestConnectorPushdownRulesWithHive.java
… plugin/trino-hive/src/test/java/io/trino/plugin/hive/orc/TestOrcPredicates.java
… plugin/trino-hive/src/test/java/io/trino/plugin/hive/parquet/TestParquetPageSourceFactory.java
… plugin/trino-hive/src/test/java/io/trino/plugin/hive/parquet/TestPruneTypeForPath.java
… plugin/trino-hive/src/test/java/io/trino/plugin/hive/parquet/predicate/TestParquetPredicateUtils.java